backend/requirements.txt
Flask>=2.0.0
Flask-CORS>=3.0.0
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.0.0
python-dotenv>=0.19.0
pymongo
firebase-admin
google-cloud-storage
google-cloud-firestore
requests
dnspython